30.5. Filter Autowiring
Back in the filter, I'm going to be tricky and add this as the second argument - bool $useLike = false - to prove that order don't matter
Each argument is passed via autowiring or by the named arguments matching.
Put your cursor back on the argument, hit Alt+Enter and go to "Initialize properties" to create the $useLike property and set it
And... that's it! I'm going to stop right there and not actually use the  property to change the query... because that would be pretty boring.
As long as it doesn't explode, we know it's working. Refresh and... yes! If we add ?search=cube... that still works too!
Next, let's talk about something totally different: API Platform's super-cool output DTO system: a middle-ground between adding custom properties to your entity and creating totally custom API resource classes.