32.1.DTO Data Transformer
To get an output DTO class to... ya know... actually work, we need to write code that converts our CheeseListing object into a CheeseListingOutput object so that API Platform can serialize it. The "thing" that does that is called a data transformer.

Creating the Data Transformer Class
Let's create one in the src/ directory: add a new directory called DataTransformer for organization and a class inside called CheeseListingOutputDataTransformer
As usual with a class that hooks into part of API Platform, this needs to implement an interface. In this situation, it's - surprise! - DataTransformerInterface!
Inside the class, go to "Code"->"Generate" - or Command+N on a Mac - and select "Implement Methods" to generate the two we need

The Data Transformer System
So here's how this works: the data transformer system exists entirely to support the input and output DTO classes that we're working with. Whenever API Platform is about to serialize an object, it checks to see if that resource has an output DTO - which we configured in CheeseListing's @ApiResource annotation. If it does, it loops over every data transformer in the system and calls supportsTransformation().
It basically asks each one:

  Hey! Apparently I need to transform a CheeseListing into a CheeseListingOutput. Do... you know how to do that?
And thanks to auto-configuration, because our new class implements DataTransformerInterface, it's instantly part of that system. In other words, our supportsTransformation() method should now be called!

supportsTransformation()
To prove it is, lets dd($data) and $to
Now, move over and refresh the endpoint. There it is! API Platform is passing us the CheeseListing and for the  argument, it's asking:

  Do you know how to convert this CheeseListing into CheeseListingOutput?
And we do!