32.2.DTO Data Transformer
For supportsTransformation(), return  instanceof CheeseListing and $to === CheeseListingOutput::class
That second part might seem unnecessary... since, in our app, a CheeseListing will always have CheeseListingOutput as its output class. But technically, you can configure a different output class on an operation-by-operation basis. So, we're checking it to be safe.
As soon as one of the data transformers returns true from supportsTransformation(), API Platform will call transform() so that we can do our work. To make sure that's happening, dd($object) and $to
When we move over and refresh... yes! It dumps the exact same thing.