32.4.DTO Data Transformer
Normalization Groups Still Apply to DTO's!
But before that, my second question is: why don't we see the title field? That has a simpler answer: normalization groups.
On CheeseListing, we set a normalizationContext with groups set to cheese:read
Thanks to the output DTO, what's actually being serialized now is a CheeseListingOutput object. But, the normalization groups still apply.
In other words, in CheeseListingOutput, we need to add that group above any properties that we want to serialize. Above title, say @Groups(), go copy the group name, and paste it here: cheese:read
Now when we try it... sweet! We have a title field!
Next: let's add the rest of the properties we need to CheeseListingOutput and see how all of this looks in our documentation. Because... similar to DailyStats, since this is not an entity, we're going to need to do a bit more work to help API Platform understand the types of each property.