33.2.Output Properties & Metadata
Adding More Fields
Ok, let's add the rest of the fields we need to this class. Check out CheeseListing: it looks like description is usually serialized and so is price
Copy the title property, paste, rename it to description... and remove the docs. Copy this and make one more property called price, which is an int
Now that we've added these properties, we need to go into our data transformer and set them. So, $output->description = $cheeseListing->getDescription() and $output->price = $cheeseListing->getPrice()
These data transformer classes are delightfully boring.
Before we try this, let's grab a couple other fields from CheeseListing. Search for cheese:read. But ignore owner for now: we'll come back to that in a minute
Ok: we also output a shortDescription field via this getShortDescription() method. Copy that whole thing and, in CheeseListingOutput paste it at the bottom
That will work exactly like before: it's referencing the description property and it has the group on it.
Back in CheeseListing, if you search again, there is one more field to move: createdAtAgo. Copy this method... then paste at the bottom. PhpStorm politely asks me if I want to import the Carbon use statement. I do!
But, hmm: this method references a  property... which we do not have inside this class. We need to add it. Add a public , but I'm not going to put any groups above this because this isn't a field that we will expose in our API directly. We just need its data:
Oh, and, by the way, we could simplify this by, instead, creating a  property, exposing that, then setting the string onto that property from our data transformer. I won't do that now, but... it's a pretty great idea and shows off the power of data transformers: you can do the work there and then have super simple DTO classes.
Anyways, back in the data transformer, set this property: $output->createdAt = $cheeseListing->getCreatedAt()
I think we're ready! Let's first refresh the documentation: open the item operation, go to schema and... yes! It did rebuild the cache that time and we can see all our custom fields and their types.
And if we go over and refresh the actual endpoint... that works to! How awesome is that? We have 5 fields and you can quickly look at our output class to know what they will be.
Next: the one field that we aren't exposing yet is the owner field. Let's add that. Though, when we do, there's going to be a slight problem.