34.1.DTO Quirks
The last field that we're missing on CheeseListingOutput is owner
No worries: in CheeseListingOutput, add public $owner. Then copy the PHPDoc from price and paste that here. We know that this will be a User object and we'll put it in the cheese:read group
Over in the data transformer, populate that with $output->owner = $cheeseListing->getOwner()
Easy enough! Try it: find your browser, refresh, and it works! The owner field is an embedded object because the phoneNumber field is being exposed.
It turns out, this detail is important. Go into the User class and look at the phoneNumber property. This is actually in two groups: owner:read an admin:read:
 290 lines  src/Entity/User.php
class User implements UserInterface
{
    /**
     * @Groups({"admin:read", "owner:read", "user:write"})
     */
    private $phoneNumber;
}

Right now, I'm logged in as an admin... and we created special code in the last tutorial to always add the admin:read group in this situation. This is the reason why we're able to see the phoneNumber on every user.
Let's see what happens when we are not an admin. Open a new tab, go to the homepage and click log out. Perfect.

User Serialization Has Changed?
Now that we're anonymous, refresh the same endpoint. Error? Interesting:
The return value of UserNormalizer::normalize() - that's a class we created in a previous tutorial - must be type array, string returned.
Let's go check that out: src/Serializer/Normalizer/UserNormalizer.php. The purpose of this is to add an extra owner:read group if the User that's being serialized is the currently-authenticated user:

 67 lines  src/Serializer/Normalizer/UserNormalizer.php
class UserNormalizer implements ContextAwareNormalizerInterface, CacheableSupportsMethodInterface, NormalizerAwareInterface
{
    public function normalize($object, $format = null, array $context = array()): array
    {
        $isOwner = $this->userIsOwner($object);
        if ($isOwner) {
            $context['groups'][] = 'owner:read';
        }
        $context[self::ALREADY_CALLED] = true;
        return $this->normalizer->normalize($object, $format, $context);
    }
}
The error says that this method is returning a string but it's supposed to return an array. And... it's right. Look at my normalize() method: I gave it an array return type:

 67 lines  src/Serializer/Normalizer/UserNormalizer.php
class UserNormalizer implements ContextAwareNormalizerInterface, CacheableSupportsMethodInterface, NormalizerAwareInterface
{
    public function normalize($object, $format = null, array $context = array()): array
    {
    }
}
But apparently, when we call $this->normalizer->normalize(), this returns a string:

 67 lines  src/Serializer/Normalizer/UserNormalizer.php
class UserNormalizer implements ContextAwareNormalizerInterface, CacheableSupportsMethodInterface, NormalizerAwareInterface
{
    public function normalize($object, $format = null, array $context = array()): array
    {
        return $this->normalizer->normalize($object, $format, $context);
    }
}
And, hmm... that makes sense. Now that we're anonymous, the phoneNumber field will not be returned. And so, when the embedded User object is serialized, instead of returning an array of fields, it is now normalized into its IRI string.