34.2.DTO Quirks
Ok! So if you normalize a User object, sometimes it will be an object and sometimes it will be an IRI string. The fix is to remove the array return type.
That was actually never needed... it's not on the normalize() method's interface. I added it simply because I thought this would always return an array.
When we refresh now... yep! The owner property is an IRI string.
But... wait a second. Why didn't we have this error before? Before we started working with the output DTO stuff... shouldn't we have had this same problem with UserNormalizer? Why wasn't it a problem until now?

DTO's Serialize Differently
Here's the answer, and it's important. When you use an output class like we're doing for CheeseListing, the object that's ultimately serialized is CheeseListingOutput:

 223 lines  src/Entity/CheeseListing.php
/**
 * @ApiResource(
 *     output=CheeseListingOutput::CLASS,
 * )
 */
class CheeseListing
{
}
And because that class isn't technically an API Resource class, it's serialized in a slightly different way internally. For the serialization nerds out there, API resource classes are usually normalized using ItemNormalizer which extends AbstractItemNormalizer. But with a DTO object, it instead uses the simpler ObjectNormalizer.

Where Did @type Go?
This causes small, but important differences. For example, when you use an output DTO, the @type field is gone. We have @id... but not @type. This actually makes one of our tests fail.
Find your terminal and run:

  symfony php bin/phpunit
Yep! One failure because one test is looking for @type. Let's open this test up: tests/Functional/CheeseListingResourceTest.php and then scroll down to testGetCheeseListingCollection(). Let's see... here it is:

 199 lines  tests/Functional/CheeseListingResourceTest.php
class CheeseListingResourceTest extends CustomApiTestCase
{
    public function testGetCheeseListingCollection()
    {
        $this->assertJsonContains(['hydra:member' => [
            0 => [
                '@type' => 'cheese',
            ]
        ]]);
    }
}
That @type is no longer being returned. For now, just delete it so that the tests will pass.
But good news! Thanks to the API Platform team, this bug has been fixed and should be released in API Platform 2.5.8. But since that hasn't been released yet at the time of this recording, we'll move on.

Run the tests again:

symfony php bin/phpunit
And... green! Next, I want to look a bit deeper at how the serialization of embedded objects is different with DTO classes and what to do about it.