35.1.DTO Quirks: Embedded Objects
We know that if none of the fields on a related object will be serialized, then API Platform should return an array of IRI strings instead of embedding the objects.

readableLink on Embedded Objects
This is a bug in how the readableLink for properties is calculated when you have a DTO. I've actually fixed this bug... but I need to finish that pull request.
Specifically, in the User class, if we search for getPublishedCheeseListings(), this is the method that gives us the cheeseListings property:

 290 lines  src/Entity/User.php
class User implements UserInterface
{
    /**
     * @SerializedName("cheeseListings")
     */
    public function getPublishedCheeseListings(): Collection
    {
        return $this->cheeseListings->filter(function(CheeseListing $cheeseListing) {
            return $cheeseListing->getIsPublished();
        });
    }
}
But because CheeseListing uses a DTO, it doesn't calculate readableLink correctly. Remember: readableLink is calculated by checking to see if the embedded object - CheeseListing has any properties that are in the same normalization groups as User. But... since CheeseListing isn't actually the object that will ultimately be serialized... API Platform should really check to see if CheeseListingOutput has any fields in the user:read group.
Anyways, one way to fix this is just to force it. We can say @ApiProperty with readableLink=false
Now, when we move over and refresh... that will force it to use IRI strings. So... this is another quirk to be aware of, but hopefully it will get fixed soon.

IRI String Problem with Multiple Output Classes
By the way, the problem of an object being embedded when it should be an IRI string gets a bit worse if you use multiple output classes. Like, if User also had a UserOutput with a cheeseListings field... even adding readableLink=false wouldn't help. If you have this situation, you can check out a conversation about it in the comments.