35.3.DTO Quirks: Embedded Objects
Re-add the @ApiProperty but this time say readableLink=true because we do want to force an embedded object
When we refresh now... yes! It's back to an embedded object. Also try /api/cheeses.jsonld... that looks good, and let's run the tests one last time:

symfony php bin/phpunit
They do pass. With output DTO's, you need to be a bit more careful, though some - but not all - of these "quirks" have already been fixed or will be soon. The important thing to keep in mind is that DTO's are not serialized in exactly the same way as ApiResource classes. So code carefully.
Next: let's talk about using an input DTO.