36.3.Input DTO Class
I'm missing a bundle that adds the integration between the dump() function and the profiler. To install it, find your terminal and run:

  composer require symfony/debug-bundle
Once this finishes... we should be able to go back to the browser, hit Execute again and... perfect! A 400 error JSON response, but no HTML.
To see the dumped variables, go down to the web debug toolbar and open the last request's profiler in a new tab. Nice! It automatically took me to the Debug section: here are the $data, $to and $context variables.
Next: let's use this information to finish our data transformer and get this thing working!