37.2.Input DTO: Denormalizing IRI Strings
Let's rename  to $input and add some PHPDoc: we know this will be a CheeseListingInput: the result of deserializing the JSON
So... now our job is simple, right? Take this CheeseListingInput object, move the data over onto a CheeseListing and return it. What could go wrong?

Denormalizing IRIs into Objects
But... wait a second: check out the dump again in the profiler. The JSON has been deserialized into this CheeseListingInput object, which is cool, but look at the owner field. It's a string! I mean... that makes sense... because we are sending this string in our JSON. But... does that mean we're supposed to query for the User object manually using this IRI?
Nope! One of the jobs of the deserializer is to take each piece of data in the JSON that we're sending and figure out what type it should be, like a string DateTime object or User object. Then it does whatever work is needed to change the raw data into that type - like creating a DateTime object from a date string. There are various normalizers that are good at doing this.
And... fortunately API Platform comes with an ItemNormalizer whose job is to change IRI strings into objects by querying the database... or... more accurately, by calling the item data provider.
So why isn't that happening here? Why isn't the IRI string being changed into a User object? Check out the CheeseListingInput class and look at the owner property
The deserializer has no idea that this property is supposed to hold a User object! And so, that ItemNormalizer doesn't know it should do its work!
Let's help it: above this add @var User
Head back to the browser, hit Execute and... once that finishes, go to the Profiler, hit Latest and tada! The owner is now a User object! Adding proper types to your properties is very important for deserialization.
Next, there's a small bug in the documentation that we actually just found a way to work around. Let's see what it is and finish our data transformer.